Move prod types to EssentialFeed production module
