+++ EssentialFeed/EssentialFeed/FeedPresentation/FeedPresenter.swift
@@ -1,0 +1,83 @@
+//
+//  FeedPresenter.swift
+//  EssentialFeed
+//
+//  Created by Radosław Serek on 30/05/2023.
+//
+
+import Foundation
+
+public protocol FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel)
+}
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
+
+public protocol FeedView {
+    func display(_ viewModel: FeedViewModel)
+}
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
+
+public protocol FeedErrorView {
+    func display(_ viewModel: FeedErrorViewModel)
+}
+
+public struct FeedErrorViewModel {
+    public let message: String?
+    
+    static var noError: Self {
+        return .init(message: nil)
+    }
+    
+    static func error(message: String) -> Self {
+        return .init(message: message)
+    }
+}
+
+public final class FeedPresenter {
+    private let loadingView: FeedLoadingView
+    private let feedView: FeedView
+    private let errorView: FeedErrorView
+
+    public static var title: String {
+        return NSLocalizedString("feed_view_title",
+                                 tableName: "Feed",
+                                 bundle: Bundle(for: FeedPresenter.self),
+                                 comment: "Title for the feed view")
+    }
+
+    private var feedLoadError: String {
+        return NSLocalizedString(
+            "feed_view_connection_error",
+            tableName: "Feed",
+            bundle: Bundle(for: FeedPresenter.self),
+            comment: "Error message displayed when we can't load the image feed from the server"
+        )
+    }
+
+    public init(loadingView: FeedLoadingView, feedView: FeedView, errorView: FeedErrorView) {
+        self.loadingView = loadingView
+        self.feedView = feedView
+        self.errorView = errorView
+    }
+    
+    public func didStartLoadingFeed() {
+        errorView.display(.noError)
+        loadingView.display(FeedLoadingViewModel(isLoading: true))
+    }
+    
+    public func didFinishLoading(with feed: [FeedImage]) {
+        feedView.display(FeedViewModel(feed: feed))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+    
+    public func didFinishLoading(with error: Error) {
+        errorView.display(.error(message: feedLoadError))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+}
