+++ EssentialFeed/EssentialFeedTests/FeedPresentation/FeedPresenterTests.swift
@@ -8,81 +8,6 @@
 import EssentialFeed
 import XCTest
 
-protocol FeedLoadingView {
-    func display(_ viewModel: FeedLoadingViewModel)
-}
-
-struct FeedLoadingViewModel {
-    let isLoading: Bool
-}
-
-protocol FeedView {
-    func display(_ viewModel: FeedViewModel)
-}
-
-struct FeedViewModel {
-    let feed: [FeedImage]
-}
-
-protocol FeedErrorView {
-    func display(_ viewModel: FeedErrorViewModel)
-}
-
-struct FeedErrorViewModel {
-    let message: String?
-    
-    static var noError: Self {
-        return .init(message: nil)
-    }
-    
-    static func error(message: String) -> Self {
-        return .init(message: message)
-    }
-}
-
-final class FeedPresenter {
-    private let loadingView: FeedLoadingView
-    private let feedView: FeedView
-    private let errorView: FeedErrorView
-
-    static var title: String {
-        return NSLocalizedString("feed_view_title",
-                                 tableName: "Feed",
-                                 bundle: Bundle(for: FeedPresenter.self),
-                                 comment: "Title for the feed view")
-    }
-
-    private var feedLoadError: String {
-        return NSLocalizedString(
-            "feed_view_connection_error",
-            tableName: "Feed",
-            bundle: Bundle(for: FeedPresenter.self),
-            comment: "Error message displayed when we can't load the image feed from the server"
-        )
-    }
-
-    init(loadingView: FeedLoadingView, feedView: FeedView, errorView: FeedErrorView) {
-        self.loadingView = loadingView
-        self.feedView = feedView
-        self.errorView = errorView
-    }
-    
-    func didStartLoadingFeed() {
-        errorView.display(.noError)
-        loadingView.display(FeedLoadingViewModel(isLoading: true))
-    }
-    
-    func didFinishLoading(with feed: [FeedImage]) {
-        feedView.display(FeedViewModel(feed: feed))
-        loadingView.display(FeedLoadingViewModel(isLoading: false))
-    }
-    
-    func didFinishLoading(with error: Error) {
-        errorView.display(.error(message: feedLoadError))
-        loadingView.display(FeedLoadingViewModel(isLoading: false))
-    }
-}
-
 final class FeedPresenterTests: XCTestCase {
     func test_title_isLocalized() {
         XCTAssertEqual(FeedPresenter.title, localized("feed_view_title"))
